+++ src/applications/vaos/tests/covid-19-vaccine/components/NewBookingPage.unit.spec.js
@@ -10,8 +10,8 @@ import { createTestStore, renderWithStoreAndRouter } from '../../mocks/setup';
 import { NewBookingSection } from '../../../covid-19-vaccine';
 import { mockFacilitiesFetchByVersion } from '../../mocks/fetch';
 import { createMockFacilityByVersion } from '../../mocks/data';
-import { mockSchedulingConfigurations } from '../../mocks/helpers.v2';
-import { getSchedulingConfigurationMock } from '../../mocks/v2';
+import { mockSchedulingConfigurations } from '../../mocks/helpers';
+import { getSchedulingConfigurationMock } from '../../mocks/mock';
 
 const initialState = {
   featureToggles: {
