+++ src/applications/vaos/tests/new-appointment/components/ReviewPage/index.va-request.unit.spec.js
@@ -19,7 +19,7 @@ import {
 } from '../../../mocks/setup';
 
 import ReviewPage from '../../../../new-appointment/components/ReviewPage';
-import { mockAppointmentSubmitV2 } from '../../../mocks/helpers.v2';
+import { mockAppointmentSubmitV2 } from '../../../mocks/helpers';
 import { createMockCheyenneFacilityByVersion } from '../../../mocks/data';
 import { mockFacilityFetchByVersion } from '../../../mocks/fetch';
